dropzone/requirements.txt
requests==2.22.0
pyftpdlib==1.5.6
pycryptodome==3.9.9
